refactor; move download into its own app
