+++ cognite_toolkit/_cdf_tk/feature_flags.py
@@ -54,10 +54,6 @@ class Flags(Enum):
         "visible": True,
         "description": "Enables the support for external libraries in the config file",
     }
-    DOWNLOAD: ClassVar[dict[str, Any]] = {  # type: ignore[misc]
-        "visible": True,
-        "description": "Enables the support for the download command",
-    }
 
     def is_enabled(self) -> bool:
         return FeatureFlag.is_enabled(self)
